requirements.txt
genai
python-dotenv
google-generativeai
md2pdf
PyPDF2